update dependency notation for correct limits
As pointed out in [this](https://github.com/astral-sh/uv/issues/11977)
uv issue, the `^` notation sets an upper bound at < next major version,
so as used, the depencies are declared as django>=3.0,<4.0.

This commit updates the dependency declaration to match the intent of
the library, as expressed in it's classifiers and readme.